pyproject.toml
[tool.poetry]
name        = "assessors"
version     = "0.1.0"
description = ""
authors     = ["Wout Schellaert <wout.schellaert@gmail.com>"]

[tool.poetry.dependencies]
python              = "^3.9"
tensorflow-gpu      = "^2.6.0"
tensorflow-datasets = "^4.4.0"
tqdm = "^4.62.3"

[tool.poetry.dev-dependencies]
ipywidgets         = "^7.6.3"
click              = "^8.0.1"
mypy               = "^0.910"
autopep8           = "^1.5.7"
pylint             = "^2.9.6"
jupyterlab         = "^3.1.1"
ipykernel          = "^6.0.3"
nbdime             = "^3.1.0"
poethepoet         = "^0.10.0"
wandb              = "^0.11.2"
Pillow             = "^8.3.1"  # Needed for wandb Image logging
click-option-group = "^0.5.3"

[build-system]
requires      = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# https://github.com/nat-n/poethepoet
[tool.poe.tasks]
jpt = { shell = "jupyter lab --no-browser --port=8888" }
run = { script = "assessors.main:run" }
gpu = { shell = "python scripts/gpu_test.py" }
